requirements.txt
httplib2==0.8
pyserial==2.6
simplejson==3.3.0